% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linker_functions.R
\name{codonMap}
\alias{codonMap}
\title{Convert IUPAC degenerate base codes to the full set of encoded bases}
\usage{
codonMap(deg_codons, parent = NA)
}
\arguments{
\item{deg_codons}{List of characters indicating the bases used at each 
position including degenerate bases that encode multiple possible nucleotides. 
For example S corresponds to G or C.}

\item{parent}{Character string of the parent sequence used for cloning 
library. This argument allows for a single sequence not included in 
degenerate library to be included in processing steps in the event there is 
carryover of a single sequence during cloning process.}
}
\value{
Dataframe containing single column with characters for all sequences 
encoded by degenerate bases.
}
\description{
This function is used internally to generate the rows of count matrix for
linker library sort-seq experiments. Each row will correspond to a unique
sequence encoded by the degenerate bases provided in \code{deg_codons} 
argument.
}
\examples{
\dontrun{
deg_codons <- c(rep(c("S", "N", "A"), 2), rep(c("V", "S", "T"), 2))
codonMap(deg_codons = deg_codons, parent = NA)
}
}
\keyword{internal}
