% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linker_functions.R
\name{countReads}
\alias{countReads}
\title{Count the number of times each unique sequence is present in each fastq file.}
\usage{
countReads(fl, deg_codons, parent = NA)
}
\arguments{
\item{fl}{Filename indicating the location of sequencing reads. For 
paired-end reads only the first (R1) file needs to be included.}

\item{deg_codons}{List of characters indicating the bases used at each 
position of sequence. Only sequences encoded by the provided bases including
degenerate bases will be counted.}

\item{parent}{Character string of the parent sequence used for cloning 
library. This argument allows for a single sequence not included in 
degenerate library to be included in processing steps in the event there is 
carryover of a single sequence during cloning process.}
}
\value{
No return. Writes files containing read counts for each sequence in
each file that can later be concatenated using read.Linker.Counts.
}
\description{
This function is used internally to count the number of times each unique
sequence is present in the trimmed read files.
}
\examples{
\dontrun{
fl <- "inst/extdata/PyronicSF_example.fastq.gz"
deg_codons <- c(rep(c("S", "N", "A"), 2), rep(c("V", "S", "T"), 2))
countReads(fl = fl, deg_codons = deg_codons)
}
}
\keyword{internal}
