% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLE_functions.R
\name{lnorm_mean}
\alias{lnorm_mean}
\title{Convert log-normal parameters μ and σ to the log-normal mean}
\usage{
lnorm_mean(x)
}
\arguments{
\item{x}{list of log-normal parameters}
}
\value{
integer idicating log-normal mean
}
\description{
Convert log-normal parameters μ and σ to the log-normal mean
}
\keyword{internal}
