% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLE_functions.R
\name{lnorm_mle}
\alias{lnorm_mle}
\title{Maximum-likelihood estimates for binned sort-seq using 
log-normal distribution}
\usage{
lnorm_mle(ri, mu, sig, gates, dj, eps)
}
\arguments{
\item{ri}{read counts for variant i}

\item{mu}{mean of log fluorescence distribution}

\item{sig}{standard deviation of log fluorescence distribution}

\item{dj}{proportionality factor for each bin describing the total number 
of cells sorted in bin j to the total number of read counts in bin j.}

\item{eps}{error rates for sorting}
}
\value{
Dataframe containing maximum likelihood estimates of mu and sigma for each
 variant i.
}
\description{
Maximum-likelihood estimates for binned sort-seq using 
log-normal distribution
}
\keyword{internal}
