% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_insertion_functions.R
\name{metaFile}
\alias{metaFile}
\title{Generate a metadata file to be used for reading in DIP count files}
\usage{
metaFile(seq.length, AA.start, AA.length, fname)
}
\arguments{
\item{seq.length}{Integer corresponding to the to total number of rows in
the DIP count files. This is approximately two times the length of DNA
sequence that was transposed to account for forward and reverse insertions.}

\item{AA.start}{Integer indication the insertion site that corresponds to an
insertion of cpGFP that is forward and in-frame in the first amino acid of
ligand-binding domain.}

\item{AA.length}{Integer indicating the total number of amino acids in the
ligand-binding domain.}

\item{fname}{Filename for saving the metadata file produced as output.}
}
\value{
Dataframe containing metadata describing mapping between insertion
sites and amino acid sequence. The file will also be written out to filename
indicated by \code{fname}
}
\description{
This function is used to generate a metadata file for annotating the
insertion sites in a transposon-mediate domain-insertion library. Each
insertion site is described by the
}
\details{
metadata columns:
\itemize{
\item insertion_site_name: Character string indicating the site of insertion
and direction.
\item site: Integer indicating the number of DNA bases before the site of
cpFP insertion.
\item forward_insertion: Logical indicating whether the insertion is in the
forward direction (TRUE) or the reverse direction (FALSE).
\item in_frame: Logical indicating whether the insertion site is in-frame (TRUE)
or out-of-frame (FALSE). Only insertion after second nucleotide in codon are
result in an in-frame translation of inserted domain.
\item productive: Logical indicating whether the insertion is in the forward
direction and in-frame (TRUE) or either reversed or out-of-frame (FALSE).
\item AA: Integer indicating which amino acid in the LBD the insertion of cpFP
occurs after. All unproductive insertions are set to AA=0.
}
}
\examples{
\dontrun{
# generating metadata for MBP domain-insertion profiling experiment
metaFile(seq.length = 1120, AA.start = 8, AA.length = 370, 
fname = "data/meta.txt")
}
}
