% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_insertion_functions.R
\name{read.DIP.Counts}
\alias{read.DIP.Counts}
\title{Read domain-insertion profiling count files into a dataframe}
\usage{
read.DIP.Counts(fnames, meta.file)
}
\arguments{
\item{fnames}{List of filenames for read count tables to be read}

\item{meta.file}{Filename for metadata file describing mapping between insertion 
sites and amino acid sequence.}
}
\value{
Matrix with the first n columns corresponding to 
the n files/samples provided by \code{fnames}, rows corresponding to 
insertion sites and entries containing read counts. The last 6 columns 
provide insertion site annotations from the metadata file.
}
\description{
This function provides a means to read the count files produced by the python
dipseq package (https://github.com/SavageLab/dipseq) into a single dataframe
for sort-seq analysis.
}
\examples{
\dontrun{
read.DIP.Counts(fnames = "example.csv", meta.file = "meta.txt")
}
}
