% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linker_functions.R
\name{read.Linker.Counts}
\alias{read.Linker.Counts}
\title{Read linker counts from multiple files into a single dataframe}
\usage{
read.Linker.Counts(count.dir)
}
\arguments{
\item{count.dir}{Character indicating the directory containing linker read 
count files.}
}
\value{
Dataframe containing columns corresponding to samples, rows 
corresponding to linker sequences and entries corresponding to read counts.
}
\description{
This function is used to concatenate linker read count files for each sample
into a single dataframe for further analysis.
}
