% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLE_functions.R
\name{reads2cells}
\alias{reads2cells}
\title{Convert read counts for each variant to estimates of cells sorted containing
each variant in each bin using the proportionality constant dj}
\usage{
reads2cells(read.counts, sorted.cells)
}
\arguments{
\item{read.counts}{Dataframe with rows corresponding to variants and columns
to gates with elements indicating read counts}

\item{sorted.cells}{List of integers indicating the total number of cells 
sorted into each gate}
}
\description{
Convert read counts for each variant to estimates of cells sorted containing
each variant in each bin using the proportionality constant dj
}
