% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linker_functions.R
\name{trimReads}
\alias{trimReads}
\title{Trim the constant sequences from the sequencing reads leaving only the 
variable regions of interest.}
\usage{
trimReads(fl, leftseq, rightseq, linklen, cutadapt)
}
\arguments{
\item{fl}{Filename indicating the location of sequencing reads. For 
paired-end reads only the first (R1) file needs to be included.}

\item{leftseq}{Character indicating the constant sequence on the 5' side of 
variable sequence. N can be used to indicate any base is allowed at certain
positions.}

\item{rightseq}{Character indicating the constant sequence on the 3' side 
of variable sequence. N can be used to indicate any base is allowed at 
certain positions.}

\item{linklen}{Integer indicating the number of basepairs that comprise the
mutated linker sequence (or mutated sequence in general).}

\item{cutadapt}{Character indicating the full path to installation of the 
software cutadapt (https://cutadapt.readthedocs.io/)}
}
\value{
No return. Writes of trimmed fastq files that are deleted after use.
}
\description{
This function is used internally to process sequencing reads by removing the
constant regions in the DNA that were not mutated in order to produce only
the variable regions that are of interest. Cutadapt 
(https://cutadapt.readthedocs.io/) must be installed and the full path 
supplied.
}
\examples{
\dontrun{
fl <- "inst/extdata/naive_R1.fastq.gz"
leftseq <- "^NNNNNACGTGAGACAGAATTTTGAGCTCCTG"
rightseq <- "^NNNNNTAGACACGAGTGGCAGCATTTCGCGCCGAGAATA"
cutadapt <- "/home/jnk/software/anaconda3/bin/cutadapt"
trimReads(fl, leftseq, rightseq, linklen = 6, cutadapt)
}
}
\keyword{internal}
